+++ extension_packages/figpack_experimental/examples/example_fmri_video.py
@@ -2,7 +2,7 @@ import os
 import urllib.request
 import numpy as np
 import nibabel as nib
-import figpack_jfm.views as jv
+import figpack_experimental.views as jv
 
 
 url = "https://s3.amazonaws.com/openneuro.org/ds006661/sub-001/func/sub-001_task-main_run-01_bold.nii.gz"